pyproject.toml
[build-system]
requires = ["setuptools>=66", "wheel>=0.40"]
build-backend = "setuptools.build_meta"

[project]
name = "stickfix-bot"
version = "0.1.0"
description = "Telegram sticker tagging assistant that keeps sticker metadata searchable."
readme = "README.md"
requires-python = ">=3.14,<4"
dependencies = [
    "python-telegram-bot>=13.15,<14",
    "PyYAML>=6.0,<7",
    "coloredlogs>=15.0,<16",
    "urllib3>=1.26,<1.27",
    "six>=1.17,<1.18",
    "uv>=0.9,<1.0",
    "ruff>=0.4,<1.0",
]
[project.optional-dependencies]
db = [
    "psycopg[binary]>=3.1,<4",
    "pgvector>=0.4.1,<0.5",
    "sqlmodel>=0.0.8,<0.0.9",
]
graph = [
    "neo4j>=6.0,<7.0",
    "neo4j-driver>=5.0,<6.0",
]
dev = [
    "pytest>=7.0,<8.0",
    "pytest-cov>=4.0,<5.0",
    "pytest-bdd>=7.0,<8.0",
    "pre-commit>=3.0,<4.0",
    "hypothesis>=6.0,<7.0",
    "pyhamcrest>=2.1,<3.0",
]

[project.urls]
Homepage = "https://github.com/r8vnhill/stickfix"
Repository = "https://github.com/r8vnhill/stickfix"

[tool.setuptools]
py-modules = ["imghdr"]

[tool.setuptools.packages.find]
include = ["bot"]
exclude = ["tests", "data", "logs"]

[tool.pytest.ini_options]
norecursedirs = [
    ".git",
    ".hypothesis",
    ".pytest-tmp",
    ".pytest_cache",
    ".venv",
    "data",
    "logs",
    "venv",
    "worktmp",
]
